pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hive-agents"
version = "0.1.0"
description = "A digital Scrum team powered by AI agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.14"
authors = [
    {name = "Hive Team"}
]
keywords = ["ai", "agents", "scrum", "automation", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "openai>=1.40.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "aiofiles>=23.2.1",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    "tree-sitter>=0.21.0",
    "chromadb>=0.4.22",
    "tiktoken>=0.5.2",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
hive = "cli:app"

[tool.setuptools]
py-modules = ["cli"]
packages = ["agents", "core", "tools", "tools.rag", "core.mcp"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
